Move ObservableEvent out of netsim
